Create PhoenixTest.Static + protocols
In preparation to add LiveView tests. We create a `PhoenixTest.Static`
to encompass the logic related to deal with static pages.

We create a `PhoenixTest.Driver` protocol that `Static` and eventually
the liveview implementation have to implement.

We also extract the HTML parsing logic into `PhoenixTest.Html` and the
assertion logic into `PhoenixTest.Assertions`.