Add Live test for multiple links (raises)
What changed?
=============

We add a Live test for handling multiple links.

By default LiveViewTest raises when there are multiple links that match
the CSS selector and text filter. So, our test raises.

But this is the first time the behavior of our Static and Live
implementations diverge. The Static implementation just clicks on the
first link.

So, we have three options:

- Be okay with discrepancies like this one (at least there are helpful
  errors when in LiveView)
- Choose to do everything like LiveView (meaning we update Static to
  raise)
- Choose to do everything like Static (a little more difficult since we
  don't control LiveViewTest's implementation)

If we do raise when there are multiple links (second option above), it's
probably not a big deal, since we should be introducing a `click_link`
that accepts a CSS selector soon.