Don't always attempt to mount LiveView
What changed?
============

Our original `PhoenixTest.visit/2` helper always attempted to mount a
live view with `live/2`. We did that because that was the easiest way to
know whether the page was a live view or a static view.

But that didn't always render good error messages. For example, when
testing in a consuming application where the page was unauthorized, the
error was confusing.

Instead, we use the fact that `live/2` has a `live/1` counterpart that
takes an already-requested `conn` struct.

Thus, we always perform the `get/2` request, and then inspect the
`assigns` for the presence of a `live_module` (that's what LiveView
tests do).

Once we do that, we then call `build/1` on each of the two strategies
which builds the required struct.