Handle a Static.click_button with form
What changed?
============

We handle a simple button that submits a form. For now, we only handle
the `put` case but we should check the form's method to determine the
correct value.

We do not handle the case when the button is a submit button for a
larger form with data.