Add Live.fill_form/3
We add a Live.fill_form/3 that automatically triggers a phx-change. I'm
uncertain whether or not we'll want to allow for a different type of
combination that isn't so LiveView-esque.

For example, in other test frameworks, I would probably do something
like:

```
session
|> fill_in(form_data)
|> click_button("Submit")
```

In that flow, `fill_in` should trigger `phx-change` if one is
available, and then submit the form with the button click.

But, if we do that, people won't get an error message that `phx-change`
isn't set. Should they expect that error when they `fill_in/2`? Maybe,
maybe not.

Perhaps, the correct assertion would just check for something to change
on the page as a result of the `phx-change`, and therefore we should
_NOT_ make it an error if the `phx-change` isn't available.

Therefore, the code above could be a nice abstraction. We'll focus on
the Static versions to have a point of comparison.