Add Static.fill_form + click_button
We add a rough implementation of a submitting a static form.

There are checks for fields being present based on form data passed.

This implementation marks a departure of how Live.submit_form and
Live.fill_form work. It combines `fill_form` with a click on the submit
button. But I think this is a good approach and perhaps we should make
the `Live` implementation adjust to this type of approach.