Allow Live to use fill_form + click_button
Underneath the hood we use the `submit_form` helper, keeping track of
the active form in `fill_form`.

This gives us parity with `Static` behavior and allows a more common way
to express form submissions. It also adds validation on the submit
button being present -- something that isn't available if we just use
`submit_form`.