Allow Live.fill_form to work without `phx-change` forms
What changed?
============

Not all Live forms will have a `phx-change` attribute. Some might just
be relying on `phx-submit`.

But we don't want users of the library to know that in those situations
they _have_ to use `submit_form`. Instead, we want to make it so that
`fill_form` is smart enough to know if there's a `phx-change` or not.

That's what we do in this commit. We allow `fill_form` to be used for
forms that don't have a `phx-change`. But in order not to lose the
input validation, we add some validation to ensure the expected inputs
(based on the form data passed) are present.