assert_has/3 filters on selector AND text
What changed?
=============

Our `assert_has/3` helper was being too restrictive too early. By using
`Html.find/2`, it was raising an error if the selector matched multiple
elements -- even if the text passed would've narrowed down the query to
a single element.

Thus, we update the `assert_has/3` helper to use `Html.find/3` which
takes into account the text as a filter. But that means that
`assert_has/3` will no longer raise if there are multiple elements found
with the same selector + text.