Allow one level of nesting in forms
We add a (gross but) working implementation that allows for nested
inputs. It only handles one level of nesting for now, but we can extend
the work to recurse and handle more levels of nesting.

So, inputs like `user[name]` can now be handled by the Static side. We
should extract common validation logic and share it with the Live side.