Allow arbitrary levels of nesting in Static forms (fill_form)
We allow for arbitrary levels of nesting in Static forms.

We should still extract that logic so that Live forms can use it too
when there's no phx-change event -- and we thus have to validate form
data manually.

Finally, we add a test that ensures the error is raised when the
validation fails. We should write more of those tests to ensure we have
good error messages.