Handle select, checkbox, and textarea in Static
What changed?
=============

We update our `Static.fill_form` helper to handle selects, checkboxes,
and textareas (in addition to inputs).

While we do that, we continue to improve error messages as we hit them
while developing the library.

NOTE: for checkboxes and selects, we're not doing more advanced
handling.

For example, in checkboxes, the value is "on" by default. But the
`fill_form` could pass `true`, and then the server would receive `true`
in the test, even though the real form would likely send "on".

Similarly, in selects, we're not validating that the value passed in
`fill_form` is a valid `option`. We can do that (with a bit more work)
since we're keeping track of the select's options.