Handle selects, checkboxes, texteareas in Live.fill_form
What changed?
============

We relax the selector we're looking form in forms from `input` to allow
any element so long as it has the correct `name` on it. Perhaps we
should be looking for `input`, `select`, `input type="checkbox"` or
`textarea`, but for now this gets us to have support.

Interestingly, LiveView doesn't respect the checkbox default value of
`"on"`. Instead, it raises an error requiring the checkbox to have a
value.