Format find_one_of and form field errors
We improve the format of `find_one_of` errors and form field errors to
be a little more clear.

The `find_one_of` function returned `{field_type, text}` (e.g.
{"button", "Save"}` which wasn't a good error message.

And the form field errors mentioned the name's of the fields that were
found but not what type of element they belonged to. So, we add that to
make it a little more clear what elements we're talking about.