Improve Assertions using Query.find
Having separated `Query.find` from the `find!` (raising) version, we can
now use that in our Assertions to provide better error messages.

This commit cleans up our `assert_has` and `refute_has` logic.

It also reveals that we're handling `Query.find/3` (when we provide
text) slightly differently than `Query.find/2`. The return values are
different when multiple records are found. We hope to improve that in a
future commit and make it consistent.

Testing note
------------

We also fully specify the error messages in assertion tests. Since these
are really important, we want to make sure that the errors not only
match with a regex, but that they're exactly as we expect them.