Raise when multiple links found Static pages
What changed?
=============

We update our `Query.find/3` result to differentiate between finding one
or many elements.

Since we now do that, we can improve our assertions. But we also bring
the static and LiveView behavior up to par in a place where we had a
difference.

LiveView raises when we find multiple elements that have the same
selector and text. So, it requires users to find a unique element.

Our Static implementation would simply follow the first link (if many
were found) instead of raising. We now update the code to raise if we
find many -- matching the LiveView behavior.