Add `Query.find` tests
Now that we're settling with a better version of `Query`, it's time to add
those tests retroactively. For now, we only cover `Query.find`
functions.

The code is covered already by the more "integration" tests. But since
the API is stabilizing, it seems good to have a layer of tests at
`Query` since it's an important piece of the code (and so we can iterate
more quickly in future work).