Adds `Query.find_one_of` and `find_one_of!`
The function `find_one_of` was a private function. We promote it to a
public function (still internal to the library) along with its `!`
equivalent.

We also improve the error messages to include potential matches --
elements that match the selector but not the content.

We'll likely move the `find_submit_buttons` function up one level, since
we can now call `find_one_of!`. The knowledge of submit buttons doesn't
belong in `Query`. The `find_one_of` function is now generic enough.