Move `find_submit_buttons` out of Query
With the changes in 82492c6, `Query` now exposes a `find_one_of!` that
we should use. `find_submit_buttons` doesn't really belong in `Query`
since it's button-specific. So, we move it up one layer to the `Static`
and `Live` implementations. There's a little bit more duplication but
the abstraction seems correct.