Use Floki.raw_html/2 to format found elements
A previous commit introduced the use of `Floki.raw_html/2` to format
the content of the found elements we print in assertions. But since the
elements themselves are floki elements, we can take it one step further
and pass the full elements into `Floki.raw_html/2`. That changes
assertion text

From:

```
<h1> tag with content "Main page"
```

To:

```
<h1 id="title">
  Main page
</h1>
```

That's much more helpful for finding which elements we're matching.

We also put `Floki.raw_html/2` in `Html.raw/1` to keep all Floki parsing
in `Html`. And we always specify `pretty: true` which will multiline
html elements, even if there's only one.