Handle form redirects in Live pages (#21)
Closes #19 

What changed?
============

Our `Live` implementation doesn't handle form redirects, be they to
another LiveView with `push_navigate/2` or to a static page via
`redirect/2`.

This commit updates the implementation to follow redirects the same way
we follow redirects when we click on links.

Testing note
-----------

We also create a new describe block to encapsulate all the testing for
`fill_form/3` + `click_button/2`.