Handle form redirects in Static pages
What changed?
============

Commit 531e5e9 updated Live form submissions to follow redirects to
other LiveView and Static pages.

This commit does the same but for Static pages.

Thus, now when a `fill_form` + `click_button`, a single-field form
`click_button`, or a `submit_button` redirects, we follow the redirect
to the other page -- whether that be a LiveView or a static page.