Handle regular form submission from a LiveView page
Closes #16

What changed?
=============

Currently, we assume that if you're in a LiveView page, you're trying to
submit LiveView forms with `phx-submit`. But people can have regular
forms in LiveView pages.

This commit handles that scenario.

In order to handle that we introduce an error message when there are
neither a `phx-submit` nor an `action`. In theory, form's don't need an
`action`. The form's `action` defaults to the same page (i.e. `"/"`).

But we don't want to just fallback to the normal form if there's no
`action`. Instead we want to notify the user of an error -- we should
expect a `phx-submit` or an `action`.