Refactor Html.Form to keep more data
What changed?
============

We refactor the `Html.Form` to use consistent `build` language -- it's
not really parsing anything, and we were using create and build in
different functions.

But we also make two improvements:

- Add an "attributes" layer. It makes it more cumbersome to get the name
  of the field, but it allows us to have all the attributes in a single
  place which makes for better errors when we want to rebuild the form's
  fields with all of its attributes.

- Add a "tag" key for each field instead of using "type" so that we can
  keep the "type" for the "type" attribute (e.g. "checkbox", "text",
  etc.)

In a sense, our `Form.build/1` function now mostly transforms tuples
into maps. The only place where we don't do that (but we probably
should) is with the contents. We should recursively walk through the
contents and build them in the same way we're building the rest of the
tags.

Then, we'll be able to use that to return better errors. There's one
place in `static.ex` where we just set the content to `[]` in an error
because we don't have access to the correct contents as they were
originally presented.

Finally, this refactoring is (hopefully) the first part of making Live
and Static use the same validations for form fields. Currently, they use
different strategies.