Use Html.Form.verify_form_fields! in Live & Static
What changed?
=============

Our `Static` implementation had a much more rich validation of form
fields, largely in part because we originally relied more  heavily on
what LiveView would do.

But since `fill_form` validates form fields in both scenarios, we
extract the logic out of `Static` and into `Html.Form` so that we now
have two new functions to validate form data:

- `validate_form_data!` which validates the action and the fields
- `validate_form_fields!` which validates the fields

The first takes a `form` that is the result of `Html.Form.build` and the
second takes the form fields from that same data structure. So it seemed
like a good place to include the validation functions.