Test nested forms in LiveView (#26)
Fixes #17
Closes #25

What changed?
============

Commit 180dc0d updated our `Live` implementation of form field
validations. That fixed an issue that was happening in nested forms in
Live that was first documented in
https://github.com/germsvel/phoenix_test/issues/17.

A helpful PR https://github.com/germsvel/phoenix_test/pull/25 created a
failing test that reproduced the error, and we include a slightly
modified version of the test here.

I thought this branch would fix the issue, but the issue was completely
resolved by the work in the previously mentioned commit. So, this commit
only introduces the additional test.