Add `Form.verify_form_*!` functions tests
What changed?
=============

Commit 180dc0d extracted `Form.verify_form_data!` and
`Form.verify_form_fields!` from the `Static` implementation so that
`Live` could use the same logic.

Since we were refactoring, the code was covered in an integration way.
But in this commit, we introduce unit tests for the
`verify_form_*!`functions so we can properly cover them and modify them
with more confidence.