Handle regular form submission from Live.submit_form
What changed?
=============

In fe755de, we updated our Live `fill_form` + `click_button`
implementation to handle regular form submission in LiveView pages.
Previously, we had assumed that any form in a LiveView page would be
using a `phx-submit`.

But at the time, we didn't update the `submit_form` helper to do the
same. This commit does that.

Once we do that, it turns out that the `submit_active_form` and
`submit_form` functions look almost identical. So we refactor some of
the logic in `Live`.