Add `open_browser/1` (#31)
What changed? 
==============

This introduces a new `open_browser/1` helper to open the default 
browser to display current HTML of `session`.

The `Live` implementation mostly wraps LiveView's `open_browser/1` 
implementation. But for the `Static` implementation, we bring a lot 
of the work that was done in LiveView into `PhoenixTest`. The 
biggest difference is that we do not bring LiveView's unique integer
generation. Instead, we rely on `System.unique_integer` to help us
generate unique names for the files. 

Finally, we opt for using the name `open_browser/1` (as opposed to 
the original suggestion of `preview`) to try to align with LiveView's
existing language so it's (hopefully) more intuitive to LiveView 
users who are already accustomed to the behavior.