Remove open_browser/1. Keep open_browser/2
What changed?
=============

The `mix docs` task was giving out warnings because `open_browser/1` was
unused. We're always using `open_browser/2`, but we typically rely on
the default function (except in tests).

So, we remove `open_browser/1` and keep `open_browser/2`. We also
include a little example snippet to show how it can be used, so people
don't think they have to pass a second argument to `open_browser/2`.