Update open_with_system_cmd/1 to support WSL2
What changed?
=============

There was an [issue] and subsequent [PR] in the LiveView repo about
`open_browser/2` not working with WSL2. The issue was fixed in [this
commit].

Since we got our `open_browser` implementation for Static pages from the
LiveView implementation, we update the `open_with_system_cmd/1` function
to reflect those changes.

[issue]: https://github.com/phoenixframework/phoenix_live_view/issues/3063
[PR]: https://github.com/phoenixframework/phoenix_live_view/pull/3064
[this commit]: https://github.com/phoenixframework/phoenix_live_view/commit/ade8b4d5523634b387eff2ab81db01b4215b1790