Relax Elixir version requirement to Elixir 1.15
Elixir 1.5 came out in June 2023. See the [announcement].

We should be able to support that.

Note: we tried relaxing the Elixir version to 1.14 (since it supports
OTP 26), but `mix format` seems to have changed and CI fails with `mix
format --check-formatted`. So we won't support 1.14 unless it's
requested and someone wants to take the time to investigate why that's
happening).

[announcement]: https://elixir-lang.org/blog/2023/06/19/elixir-v1-15-0-released/