Add `PhoenixTest.assert_has/2`
What changed?
============

We add an `assert_has/2` helper that does not require the user to pass
text. Sometimes, we want to assert that an element is present on the
page even if we don't pass the text.

Recently, that was something we wanted to do in a project when we were
trying to assert the value of an input. Since the `value` is passed as
part of the CSS selector, we did not need the third argument.