Add `PhoenixTest.refute_has/2`
What changed?
=============

In commit 3756a47, we added a `PhoenixTest.assert_has/2` helper because
there are cases when we want to assert an element is present WITHOUT
specifying the text (e.g. an input where the `value=` attribute is what
the user sees).

We now introduce a similar `refute_has/2` helper that doesn't require
passing text for the same reasons.

Note: we also remove an unused pipeline in the test router.