Adds `PhoenixTest.fill_in/3`
What changed?
=============

We add a new (but, for now, undocumented) `PhoenixTest.fill_in/3`
function (inspired by Wallaby and Capybara).

The function takes a new approach towards filling out fields (text
inputs and textareas) in forms.

The goal is to rely less on the map structure like LiveViewTest does and
more on labels. Ideally, each input should be pointed to by a label.
That's helpful for screenreaders, but it also makes nicer tests to read.

Right now, people have to know how Phoenix builds the nested structure
of forms, and it ties the test to the implementation of an input's
`name`.

It also has the benefit of behaving more like a page would behave in
real life. For example, when a user partially fills a form with
`phx-change` on it, we would expect to see `phx-change` triggered as
many times as we changed the form. Now, that will automatically happen
since `fill_in` delegates to `render_change` each time the field is
filled.

Helpful `Query` functions
------------------------

This work has the added benefit of introducing two `Query` functions
that should be helpful in upcoming work.

1. `Query.find_by_label` (and the `!` version) are helpful anytime we
   want to find an element by its label. We include errors that are
   helpful guides to ensure a label has a `for` that points to an `id`.
   It even handles implicit label <> input associations (i.e. inputs
   nested within the label).

2. `Query.find_ancestor` helps us find an ancestor elements. In our
   case, we need it to find the form for a given input. But in the
   future, we could use this to better support submit buttons inside
   forms. Right now there's a bit of guesswork when a `click_button`
   comes after a `fill_form` and that's not good. There are subtle bugs
   there (e.g. a user clicks a "cancel" button after filling out a form
   in a modal -> we assume it's clicking submit).

Utils modules
------------

We include a `Utils` module to capture a function that doesn't seem to
have a good home. I'm not sure we want to keep that in the long run, but
for now, it serves a purpose.

Test notes
----------

We also add a `Selectors.input` helper for assertions but we do not
expose that yet either. We want to land at a good API before we do that.