Move `text` to option in assert_has/refute_has
What changed?
============

In order to support the `count` option when a `text` filter is passed
(and in anticipation of future ones like `at` and `exact`), we make
`text` an option:

```elixir
|> assert_has(".posts", text: "Hello world")

|> assert_has(".posts", text: "Duplicate post", count: 2)
```

Note that for now, we leave the `assert_has/2` function that takes the
text as a positional argument:

```elixir
|> assert_has(".posts", "Hello world")
```

We'll deprecate that and give people time to migrate to the `text:`
option version.

Note on error messages
-----------------------

Now that we're passing options everywhere, error messages are getting
complicated. So, we take the time to extract a few private functions
that build those messages and keep them more consistent.