Deprecate text as third arg in assertions and document options
What changed?
=============

We deprecate the `assert_has` and `refute_has` functions that take the
`text` as a positional argument.

The work done to support the deprecations was already done in other
commits. This commit simply adds a deprecation warning (we cannot use
the `@deprecate` tag because we still use a 3-arity function), and
updates the documentation to refer to `count` and `text` as options.

We also update all of our internal tests to use `text:` instead of
positional `text` to remove the deprecation warnings in our own test
suite.