Add `exact` option to assertions
What changed?
=============

By default our `assert_has` and `refute_has` helpers do a substring
match when they take in a text filter. (i.e. we use `=~` when comparing
strings)

So, the following:

```elixir
assert_has(".posts", text: "Aragorn")
```

would match any of the following elements:

```html
<div class="posts">
  Aragorn
</div>

<div class="posts">
  Aragorn son of Arathorn
</div>

<div class="posts">
  Aragorn <span class="italics">son of Arathorn</span>
</div>
```

We typically do that because we don't want the third assertion to fail.

But, sometimes we're making an assertion about an element that should
only contain text. In those cases, we want the exact match (compare
string with `==`).

So, we introduce the `exact` option.