Add check and uncheck
What changed?
============

We add `check` and `uncheck` functions to work with our new way of
interacting with forms.

Note that these rely on a form being present. They are not supposed to
be used outside of a form. We could make it so in the future, but for
now we keep them trying to replace the form, no trying to also handle
clicks (e.g. via `render_click` on `phx-click` presence).

Note on uncheck error message
----------------------------

It would be nice to add a better error when we're trying to find the
hidden input to "uncheck" but we can't. Right now it just fails with a
maybe surprising error if you don't know that checkboxes need hidden
inputs to "uncheck".

https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Input/checkbox#value