Add `PhoenixTest.choose/2`
We add `PhoenixTest.choose/2` to Live and Static. It serves to choose an
option (by label) in radio buttons.

It also (automatically) handles a default "checked" value in Live
(thanks to LiveView somehow). But we don't implement the default
"checked" value in Static yet.

We'll do that in a subsequent commit where we (ideally) can handle all
default values (including any hidden inputs).