Add 'checked' radio button default to Static
What changed?
=============

In a previous commit, we introduced the `choose` function. That function
didn't handle the case when a user doesn't select any option but the
HTML has a `checked` default.

This commit does that by "prepending" the default value when
`click_button` is called. (by "prepending" we mean adding the default
but letting a value that has been selected override it).