Use `Field.find_input!` instead of `find_radio_button!`
What changed?
============

The `find_radio_button!` and `find_input!` functions were doing
basically the same thing (finding inputs) but the latter accepts a
`value` to override (since that's what we need when a user is setting
the value).

So, we use the `find_input!` function and remove the
`find_radio_button!`.