Refactoring: Remove `private` key from sessions
What changed?
=============

Our `session` structs have a `private` key that we use to keep track of
the "active form". That `private` key is unnecessary since the entire
session is technically private.

Keeping the active form under `private` makes the code more complex for
no reason. So, this commit removes that indirection and puts an
`active_form` directly on the `Live` and `Static` "session" structs.