Reset ActiveForm correctly
What changed?
=============

There was a bug when a user clicked on an unrelated button and then
tried to use the new `fill_in` (or other new form helpers). Since
`click_button` was popping the active_form without adding a new one in
its place, we would get errors because we expect a map where there
wasn't one.

Thankfully, the `fill_in` work is still private API, so no one should be
using this. I ran into the bug testing this on my own.

This commit fixes the issue on both the Live and Static implementations
by properly resetting the ActiveForm in the session.