Include button name/value in form submissions
We include a button's  `name/value` in form submissions. For now Static
and Live have partial implementations.

In both scenarios, it only works for forms that have an active form. We
should support other forms (e.g. single button submits and data
attribute forms), but we leave that for a future commit.