Static "single button form" handles default data
What changed?
=============

We update the Static single button form to handle default data such as
hidden inputs and a "checked" radio button. We also include the name and
value of the button (if it has any).

Even though it's called the "single button form" in the Static
implementation, this is also what will happen if someone triggers a
`click_button` on a form without having done a `fill_form`. So, it also
covers the case when people want to just hit the button without clicking
`fill_form` and seeing what's included.