Update Static active form submission to use Form
What changed?
=============

We continue refactoring our button logic in Static. Since we now have a
`Form` module/struct that handles default data, we use the form when
submitting an "active form". That allows us to remove the need to check
for default values in `Static`. Instead, we combine the "default" form
data from the `Form` with the ActiveForm's form data.

Ideally, we could update the HTML when we change a form, and then we
could rely solely on `Form` instead of having to keep `ActiveForm` in
our session. But that'll take more work to get to.