Refactor Static.fill_form to use Form
What changed?
============

We refactor Static.fill_form/3 to use the new Form struct. We include it
as the `ActiveForm` (even though they're not the same data structure),
but they both have a `form_data` field.

Ideally, we will be able to remove the `active_form` data struct in
time. For now, we just try to consolidate some of the repetitive logic
behind `Form`.