Use Button in Live.click_button
What changed?
=============

We use the `Button` module/struct in `Live` so we can push more of the
parsing of data into `Button` rather than having it directly in `Live`.

Since we use `Button.find!` in `click_button`, a couple of the expected
errors change, but they only change in message not in behavior (i.e.
there's no button in those tests). So this is still a refactoring.