`Live.click_button` submits form (even without fill_form)
What changed?
============

We update the `click_button` to submit a form (with defaults such as a
checked radio button and a button's value) even if the form hasn't been
filled with `fill_form`.

In order to do that, we had to change how `Form` works. `Form` included
the button data in `form_data` but for LiveView, we need to pass the
button data separately (as event data in `render_submit`).

So, this commit includes a lot of changes.

Implementation notes
--------------------

We also had to revert the order in which we check form logic:

- first we check if it's a `phx-click` button
- if not, then we submit the form

Through all that, we also had to handle clicking on different buttons in
Live correctly.

Our previous test wasn't testing things very well (the payload). So we
update the "other" button to be clicked to be another form.

Future work
-----------

Currently, we use `Form` differently in Live and Static.

In Live, `Form` is the representation of the HTML form we've found.
What's on the page. (ideally, we'll be able to rely solely on that, but
we need to update the HTML for that to happen)

ActiveForm` is a thinner concept but that's what we keep in memory.

We don't conflate the two since one has in-memory behavior and the other
is just about representing what's on the page.

In Static, our `Form` has almost completely replaced `ActiveForm`.
That's because we use a lot of the form data much more.