Use `form.selector` to compare active-form and form
What changed?
============

Since we create composite form selectors in `Form` (unless we have an
id), it's the most unique identifier we have for a form. Thus, we use
that as a form of comparison when trying to identify if we're submitting
an active form or a different form.

In order for that to work, we update our `Form.find!` to use the
composite selector instead of just relying on whatever selector was
passed in to the `find!`