Refactor: clean up use of ActiveForm vs Form
What changed?
============

We clean up the use of ActiveForm vs Form (mostly in Static) which
allows us to remove one of the `submit` helpers.

Overall, we want to keep the responsibilities of `ActiveForm` and `Form`
separate. `ActiveForm` is the in-memory data that's being filled, and
`Form` is the data we parse from `HTML` based on a selector.