Follow multiple redirects in Live (#59)
What changed?
============

When we `click_link` we follow redirects. But in `Live`, we only follow
the redirect once. This commit updates the code so that we continue
following them until we get a mounted LiveView or an HTML response.

The `Static` implementation already followed links more than once, but
we add a test here to keep the implementation tests at the same level.