Do not include disabled inputs in default form data
What changed?
============

We update `Form`'s functions to not include disabled inputs in the
default form data. We should not include disabled inputs' values in the
payload.

From [MDN's docs]:

> Specifically, disabled inputs do not receive the click event, and
disabled inputs are not submitted with the form.

We also refactor the `find!` and `find_by_descendant!` functions to use
a `build` function to build the data structure.

[MDN's docs]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Input#disabled